backend/student-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
bcrypt==4.1.1
pyjwt==2.8.0
